requirements.txt
absl-py==1.4.0
appdirs==1.4.4
biopython==1.84
biotite==0.37.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
docker-pycreds==0.4.0
einops==0.8.0
filelock==3.9.0
fsspec==2023.4.0
ftfy==6.2.0
gitdb==4.0.10
GitPython==3.1.34
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.57.0
huggingface-hub==0.16.4
idna==3.4
iopath==0.1.10
Jinja2==3.1.2
joblib==1.3.2
Markdown==3.4.4
MarkupSafe==2.1.3
mpmath==1.2.1
msgpack==1.0.5
networkx==3.0rc1
numpy==1.25.2
oauthlib==3.2.2
packaging==23.1
pathtools==0.1.2
Pillow==9.3.0
pip==23.2.1
portalocker==2.10.1
protobuf==4.24.2
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
pytorch-triton==2.1.0+e6216047b8
PyYAML==6.0.1
regex==2024.5.15
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.3.3
scikit-learn==1.3.0
scipy==1.11.2
sentry-sdk==1.30.0
setproctitle==1.3.2
setuptools==68.0.0
six==1.16.0
smmap==5.0.0
sympy==1.11.1
tensorboard==2.14.0
tensorboard-data-server==0.7.1
threadpoolctl==3.2.0
timm==0.9.7
torch==2.1.0.dev20230904+cu121
torch-geometric==2.3.1
torch-scatter==2.1.1
torchaudio==2.2.0.dev20230904+cu121
torchvision==0.16.0.dev20230904+cu121
tqdm==4.66.1
typing_extensions==4.4.0
urllib3==1.26.16
wandb==0.15.9
wcwidth==0.2.13
Werkzeug==2.3.7
wheel==0.38.4
